spec -> pkg, test fixes
