+++ pkg/api/types/v1/slinkee.go
