+++ pkg/clientset/v1/api.go
@@ -3,7 +3,7 @@ package v1
 import (
 	"context"
 
-	v1 "github.com/vultr/slinkee/spec/api/types/v1"
+	v1 "github.com/vultr/slinkee/pkg/api/types/v1"
 
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/kubernetes/scheme"
