+++ pkg/clientset/v1/slinkee.go
@@ -3,7 +3,8 @@ package v1
 import (
 	"context"
 
-	v1 "github.com/vultr/slinkee/spec/api/types/v1"
+	v1 "github.com/vultr/slinkee/pkg/api/types/v1"
+
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/kubernetes/scheme"
