+++ pkg/connectors/errors.go
