+++ pkg/munge/constants.go
