+++ pkg/slurm/constants.go
