+++ pkg/slurm/create_slurm_conf.go
@@ -8,7 +8,8 @@ import (
 	"strconv"
 	"strings"
 
-	v1s "github.com/vultr/slinkee/spec/api/types/v1"
+	v1s "github.com/vultr/slinkee/pkg/api/types/v1"
+
 	"go.uber.org/zap"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
