+++ pkg/slurm/create_slurmrestd.go
@@ -5,16 +5,17 @@ import (
 	"fmt"
 
 	"github.com/vultr/slinkee/cmd/slinkee/config"
-	v1s "github.com/vultr/slinkee/spec/api/types/v1"
+	v1s "github.com/vultr/slinkee/pkg/api/types/v1"
 
 	"go.uber.org/zap"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/kubernetes"
 )
 
-func buildToolboxDeployment(client kubernetes.Interface, wl *v1s.Slinkee) error {
+func buildSlurmrestdDeployment(client kubernetes.Interface, wl *v1s.Slinkee) error {
 	log := zap.L().Sugar()
 
 	ds := client.AppsV1().Deployments(wl.Namespace)
@@ -25,28 +26,28 @@ func buildToolboxDeployment(client kubernetes.Interface, wl *v1s.Slinkee) error 
 	}
 
 	mungeCont := mkMungeContainer(wl)
-	slurmToolboxCont := mkSlurmToolboxContainer(wl)
+	slurmrestdCont := mkSlurmrestdContainer(wl)
 
 	log.Infof("munged container: %+v", *mungeCont)
-	log.Infof("slurm-toolbox container: %+v", *slurmToolboxCont)
+	log.Infof("slurmrestd container: %+v", *slurmrestdCont)
 
 	if aff != nil {
 		log.Infof("affinity: %+v", *aff)
 	}
 
-	slurmToolboxDep := &appsv1.Deployment{
+	slurmrestdDep := &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      fmt.Sprintf("%s-slurm-toolbox", wl.Name),
+			Name:      fmt.Sprintf("%s-slurmrestd", wl.Name),
 			Namespace: wl.Namespace,
 			Labels: map[string]string{
-				"app":                          "slurm-toolbox",
+				"app":                          fmt.Sprintf("%s-slurmrestd", wl.Name),
 				"app.kubernetes.io/managed-by": "slinkee",
 			},
 		},
 		Spec: appsv1.DeploymentSpec{
 			Selector: &metav1.LabelSelector{
 				MatchLabels: map[string]string{
-					"app": "slurm-toolbox",
+					"app": fmt.Sprintf("%s-slurmrestd", wl.Name),
 				},
 			},
 			Template: v1.PodTemplateSpec{
@@ -54,7 +55,7 @@ func buildToolboxDeployment(client kubernetes.Interface, wl *v1s.Slinkee) error 
 					Name:      wl.Name,
 					Namespace: wl.Namespace,
 					Labels: map[string]string{
-						"app":                          "slurm-toolbox",
+						"app":                          fmt.Sprintf("%s-slurmrestd", wl.Name),
 						"app.kubernetes.io/managed-by": "slinkee",
 					},
 				},
@@ -64,7 +65,7 @@ func buildToolboxDeployment(client kubernetes.Interface, wl *v1s.Slinkee) error 
 						*mungeCont,
 					},
 					Containers: []v1.Container{
-						*slurmToolboxCont,
+						*slurmrestdCont,
 					},
 					RestartPolicy:    v1.RestartPolicyAlways,
 					ImagePullSecrets: []v1.LocalObjectReference{},
@@ -101,22 +102,64 @@ func buildToolboxDeployment(client kubernetes.Interface, wl *v1s.Slinkee) error 
 		},
 	}
 
-	log.Infof("slurm_toolbox deployment: %+v", slurmToolboxDep)
+	log.Infof("slurmrestd deployment: %+v", slurmrestdDep)
 
-	_, err2 := ds.Create(context.TODO(), slurmToolboxDep, metav1.CreateOptions{})
+	_, err2 := ds.Create(context.TODO(), slurmrestdDep, metav1.CreateOptions{})
 	if err2 != nil {
 		return err2
 	}
 
-	log.Infof("slurm_toolbox deployments %s created", wl.Name)
+	log.Infof("slurmrestd deployments %s created", wl.Name)
 
 	return nil
 }
 
-func mkSlurmToolboxContainer(wl *v1s.Slinkee) *v1.Container {
+func buildSlurmrestdService(client kubernetes.Interface, wl *v1s.Slinkee) error {
+	log := zap.L().Sugar()
+
+	svc := client.CoreV1().Services(wl.Namespace)
+
+	svcSpec := &v1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      fmt.Sprintf("%s-slurmrestd", wl.Name),
+			Namespace: wl.Namespace,
+			Labels: map[string]string{
+				"app":                          fmt.Sprintf("%s-slurmrestd", wl.Name),
+				"app.kubernetes.io/managed-by": "slinkee",
+			},
+		},
+		Spec: v1.ServiceSpec{
+			Type: v1.ServiceTypeClusterIP,
+			Ports: []v1.ServicePort{
+				{
+					Name:       "slurmrestd",
+					Port:       6820,
+					Protocol:   v1.ProtocolTCP,
+					TargetPort: intstr.FromString("slurmrestd"),
+				},
+			},
+			Selector: map[string]string{
+				"app": fmt.Sprintf("%s-slurmrestd", wl.Name),
+			},
+		},
+	}
+
+	log.Infof("slurmrestd service: %+v", svcSpec)
+
+	_, err2 := svc.Create(context.TODO(), svcSpec, metav1.CreateOptions{})
+	if err2 != nil {
+		return err2
+	}
+
+	log.Infof("slurmrestd service %s created", wl.Name)
+
+	return nil
+}
+
+func mkSlurmrestdContainer(wl *v1s.Slinkee) *v1.Container {
 	c := v1.Container{
-		Name:  "slurm-toolbox",
-		Image: config.GetSlurmSlurmToolboxImage(),
+		Name:  "slurmrestd",
+		Image: config.GetSlurmSlurmrestdImage(),
 	}
 
 	c.VolumeMounts = []v1.VolumeMount{
@@ -143,13 +186,17 @@ func mkSlurmToolboxContainer(wl *v1s.Slinkee) *v1.Container {
 
 	c.Ports = []v1.ContainerPort{
 		{
-			Name:          "slurmctld",
-			ContainerPort: 6817,
+			Name:          "slurmrestd",
+			ContainerPort: 6820,
 		},
-		{
-			Name:          "slurmd",
-			ContainerPort: 6818,
-		},
+	}
+
+	// slurmrestd will NOT run as root, pull uid/gid from container for the slurm user/group
+	var runAsUser int64 = 64030
+	var runasGroup int64 = 64030
+	c.SecurityContext = &v1.SecurityContext{
+		RunAsUser:  &runAsUser,
+		RunAsGroup: &runasGroup,
 	}
 
 	return &c
