+++ pkg/slurm/delete.go
