+++ pkg/slurm/errors.go
