+++ pkg/slurm/read.go
