+++ pkg/slurm/templates.go
