+++ pkg/slurm/update.go
