+++ pkg/util/rnd/rnd.go
